backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
pymongo==4.6.1
opencv-python==4.10.0.84
numpy==2.0.0
python-dotenv==1.0.0
Pillow==10.4.0
openpyxl==3.1.2
dnspython==2.4.2

# Deep Learning Dependencies for CNN Face Recognition (PyTorch-based)
torch==2.5.1
torchvision==0.20.1
facenet-pytorch==2.6.0
scikit-learn==1.5.2
